WIP Notifications
Changes:
- Only admins or group owners can update or delete a group;
- For open groups (isOpen = false) default membership is 'member', for closed groups (isOpen = true) default membership is 'pending';
- Membership can be altered, by group owner or admin, using ChangeMembershipGroup.

+++ core/mutation.py
@@ -2,8 +2,17 @@ import graphene
 import reversion
 from django.contrib.contenttypes.models import ContentType
 from core.lib import get_id
-from .models import Comment, Group
+from .models import Comment, Group, User
 from .nodes import Node, GroupNode, CommentNode
+from django.db import transaction
+
+
+class GroupContainsMembers(Exception):
+    pass
+
+class NotAuthorized(Exception):
+    pass
+
 
 class CommentInput(graphene.InputObjectType):
     description = graphene.String(required=True)
@@ -134,6 +143,8 @@ class UpdateGroup(graphene.Mutation):
         try:
             with reversion.create_revision():
                 group = Group.objects.get(pk=get_id(id))
+                if not group.can_change(info.context.user):
+                    raise NotAuthorized
                 group.name = input['name']
                 group.description = input['description']
                 group.is_open=input['is_open']
@@ -160,7 +171,16 @@ class DeleteGroup(graphene.Mutation):
         try:
             with reversion.create_revision():
                 group = Group.objects.get(pk=get_id(id))
-                group.delete()
+                if not group.can_change(info.context.user):
+                    raise NotAuthorized
+                if group.members.exclude(user=info.context.user).exists():
+                    #other members exist, cannot delete group
+                    raise GroupContainsMembers
+                with transaction.atomic():
+                    #owner will only leave group to clear all members from group.
+                    #if group.delete fails, owner membership delete will be rollbacked
+                    group.leave(info.context.user)
+                    group.delete()
 
                 reversion.set_user(info.context.user)
                 reversion.set_comment("deleteGroup mutation")
@@ -181,9 +201,10 @@ class JoinGroup(graphene.Mutation):
     def mutate(self, info, id):
         group = Group.objects.get(pk=get_id(id))
         if group.can_join(info.context.user):
-            group.join(info.context.user)
+            group.join(info.context.user, 'member')
             ok = True
         else:
+            group.join(info.context.user, 'pending')
             ok = False
 
         return JoinGroup(ok=ok, group=group)
@@ -206,6 +227,39 @@ class LeaveGroup(graphene.Mutation):
 
         return LeaveGroup(ok=ok, group=group)
 
+class MembershipInput(graphene.InputObjectType):
+    userid = graphene.ID(required=True)
+    type = graphene.String(required=True)
+
+class ChangeMembershipGroup(graphene.Mutation):
+
+    class Arguments:
+        id = graphene.ID(required=True)
+        input = MembershipInput(required=True)
+
+    ok = graphene.Boolean()
+    group = graphene.Field(lambda: GroupNode)
+
+    def mutate(self, info, id, input):
+        try:
+            with reversion.create_revision():
+                group = Group.objects.get(pk=get_id(id))
+                if not group.can_change(info.context.user):
+                    raise NotAuthorized
+                user = User.objects.get(pk=get_id(input['userid']))
+                group.join(user, input['type'])
+                group.save()
+
+                reversion.set_user(info.context.user)
+                reversion.set_comment("updateGroup mutation")
+
+            ok = True
+        except:
+            group = None
+            ok = False
+
+        return ChangeMembershipGroup(group=group, ok=ok)
+
 class Mutation(graphene.ObjectType):
     create_comment = CreateComment.Field()
     update_comment = UpdateComment.Field()
@@ -213,5 +267,6 @@ class Mutation(graphene.ObjectType):
     create_group = CreateGroup.Field()
     update_group = UpdateGroup.Field()
     delete_group = DeleteGroup.Field()
+    change_membership_group = ChangeMembershipGroup.Field()
     join_group = JoinGroup.Field()
     leave_group = LeaveGroup.Field()
